+++ modules/hanna-utils/src/index.tests.ts
@@ -88,5 +88,7 @@ import type {
   // PickIndexed,
   // RequiredKeys,
   // OptionalKeys,
+  OmitDistributive,
+  PickDistributive,
 } from './index.js';
 /* eslint-enable @typescript-eslint/no-unused-vars, unused-imports/no-unused-imports-ts, import/first, simple-import-sort/imports */